pyproject.toml
[project]
name = "traffic-mapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "googlemaps>=4.10.0",
    "python-dotenv>=1.1.0",
]
